move GroupMap into a module
