+++ src/linear_relation/collections.rs
@@ -1,0 +1,120 @@
+//! # Collections for Group and Scalar Vars
+//!
+//! This module provides collections of group elements and scalars, [GroupMap] and [ScalarMap].
+//! These collections act as a mapping of opaque variable references to values.
+
+use alloc::format;
+use alloc::vec::Vec;
+use core::marker::PhantomData;
+use group::prime::PrimeGroup;
+
+use crate::errors::InvalidInstance;
+
+use super::GroupVar;
+
+/// Ordered mapping of [GroupVar] to group elements assignments.
+#[derive(Clone, Debug)]
+pub struct GroupMap<G>(Vec<Option<G>>);
+
+impl<G: PrimeGroup> GroupMap<G> {
+    /// Assign a group element value to a point variable.
+    ///
+    /// # Parameters
+    ///
+    /// - `var`: The variable to assign.
+    /// - `element`: The value to assign to the variable.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the given assignment conflicts with the existing assignment.
+    pub fn assign_element(&mut self, var: GroupVar<G>, element: G) {
+        if self.0.len() <= var.0 {
+            self.0.resize(var.0 + 1, None);
+        } else if let Some(assignment) = self.0[var.0] {
+            assert_eq!(
+                assignment, element,
+                "conflicting assignments for var {var:?}"
+            )
+        }
+        self.0[var.0] = Some(element);
+    }
+
+    /// Assigns specific group elements to point variables (indices).
+    ///
+    /// # Parameters
+    ///
+    /// - `assignments`: A collection of `(GroupVar, GroupElement)` pairs that can be iterated over.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the collection contains two conflicting assignments for the same variable.
+    pub fn assign_elements(&mut self, assignments: impl IntoIterator<Item = (GroupVar<G>, G)>) {
+        for (var, elem) in assignments.into_iter() {
+            self.assign_element(var, elem);
+        }
+    }
+
+    /// Get the element value assigned to the given point var.
+    ///
+    /// Returns [`InvalidInstance`] if a value is not assigned.
+    pub fn get(&self, var: GroupVar<G>) -> Result<G, InvalidInstance> {
+        match self.0.get(var.0) {
+            Some(Some(elem)) => Ok(*elem),
+            Some(None) | None => Err(InvalidInstance::new(format!(
+                "unassigned group variable {}",
+                var.0
+            ))),
+        }
+    }
+
+    /// Iterate over the assigned variable and group element pairs in this mapping.
+    // NOTE: Not implemented as `IntoIterator` for now because doing so requires explicitly
+    // defining an iterator type, See https://github.com/rust-lang/rust/issues/63063
+    #[allow(clippy::should_implement_trait)]
+    pub fn into_iter(self) -> impl Iterator<Item = (GroupVar<G>, Option<G>)> {
+        self.0
+            .into_iter()
+            .enumerate()
+            .map(|(i, x)| (GroupVar(i, PhantomData), x))
+    }
+
+    pub fn iter(&self) -> impl Iterator<Item = (GroupVar<G>, Option<&G>)> {
+        self.0
+            .iter()
+            .enumerate()
+            .map(|(i, opt)| (GroupVar(i, PhantomData), opt.as_ref()))
+    }
+
+    /// Add a new group element to the map and return its variable index
+    pub fn push(&mut self, element: G) -> GroupVar<G> {
+        let index = self.0.len();
+        self.0.push(Some(element));
+        GroupVar(index, PhantomData)
+    }
+
+    /// Get the number of elements in the map
+    pub fn len(&self) -> usize {
+        self.0.len()
+    }
+
+    /// Check if the map is empty
+    pub fn is_empty(&self) -> bool {
+        self.0.is_empty()
+    }
+}
+
+impl<G> Default for GroupMap<G> {
+    fn default() -> Self {
+        Self(Vec::default())
+    }
+}
+
+impl<G: PrimeGroup> FromIterator<(GroupVar<G>, G)> for GroupMap<G> {
+    fn from_iter<T: IntoIterator<Item = (GroupVar<G>, G)>>(iter: T) -> Self {
+        iter.into_iter()
+            .fold(Self::default(), |mut instance, (var, val)| {
+                instance.assign_element(var, val);
+                instance
+            })
+    }
+}
